# contracts/core-sim.openapi.yaml
# The contract of the simulated surrounding insurance systems, and the single source of truth for
# them. Three things follow from this document and nothing here is written by hand a second time:
#
#   * src/claim_triage/contract/  — the wire models and the typed client, generated by
#     `uv run poe generate` (src/claim_triage/codegen.py);
#   * src/claim_triage/core_sim/  — the service that answers this contract, validating every request
#     and every response with those generated models;
#   * tests/contract/             — the contract tests, which drive the generated client against the
#     running service and fail when the service and this document disagree.
#
# Conventions this document uses deliberately:
#
#   * Money is a decimal *string* — `format: decimal` with `x-max-digits` and `x-decimal-places` —
#     because a JSON number cannot carry a fixed scale without rounding through a float. The client
#     and the service both hold it as `Decimal`.
#   * Every operation declares its `operationId`, which is the generated client's method name.
#   * Every write requires an `Idempotency-Key` and is deduplicated by it (see the header's own
#     description for the semantics). Replaying a key returns the response the first call recorded;
#     reusing it for a different request is a documented 409.
#   * Every failure is an `ErrorResponse` with a `code` from one enum, so the generated client maps
#     each documented failure onto its own typed exception.
openapi: 3.1.0
info:
  title: Simulated surrounding insurance systems
  version: "0.1.0"
  description: |
    The insurance systems the triage pipeline integrates with: policy lookup, claim lookup, claim
    creation, document attachment, status update and party history. They are simulated by
    `core-sim` (see docs/adr/0001) and hold no real policy or personal data.
servers:
  - url: http://localhost:8080
    description: The development stack's published port (compose.yaml)
paths:
  /healthz:
    get:
      operationId: read_readiness
      summary: Report whether the systems are ready
      responses:
        "200":
          description: The systems are ready and the claims they hold are reachable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Readiness"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /policies/{policy_number}:
    get:
      operationId: read_policy
      summary: Look a policy up as the system of record holds it
      parameters:
        - name: policy_number
          in: path
          required: true
          description: The policy number the claim is made against.
          schema:
            type: string
            minLength: 1
            maxLength: 64
      responses:
        "200":
          description: The policy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
        "404":
          $ref: "#/components/responses/PolicyNotFound"
        "422":
          $ref: "#/components/responses/InvalidPayload"
  /claims:
    post:
      operationId: create_claim
      summary: Take a claim into the systems
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimSubmission"
      responses:
        "201":
          description: The claim as the systems received it.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
        "404":
          $ref: "#/components/responses/PolicyNotFound"
        "409":
          $ref: "#/components/responses/IdempotencyKeyReuse"
        "422":
          $ref: "#/components/responses/InvalidPayload"
  /claims/{claim_id}:
    get:
      operationId: read_claim
      summary: Read a claim back
      parameters:
        - $ref: "#/components/parameters/ClaimId"
      responses:
        "200":
          description: The claim as the systems hold it.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
        "404":
          $ref: "#/components/responses/ClaimNotFound"
        "422":
          $ref: "#/components/responses/InvalidPayload"
  /claims/{claim_id}/status:
    post:
      operationId: record_claim_status
      summary: Record where the pipeline got to with a claim
      parameters:
        - $ref: "#/components/parameters/ClaimId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimStatusUpdate"
      responses:
        "200":
          description: The claim with this transition appended to its history.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
        "404":
          $ref: "#/components/responses/ClaimNotFound"
        "409":
          $ref: "#/components/responses/IdempotencyKeyReuse"
        "422":
          $ref: "#/components/responses/InvalidPayload"
  /claims/{claim_id}/documents:
    post:
      operationId: attach_claim_document
      summary: Attach a document the pipeline read out of a claim
      description: |
        The bytes stay with the pipeline; what the surrounding systems hold is which document a
        claim carries, what the pipeline classified it as, and under which filename it arrived.
      parameters:
        - $ref: "#/components/parameters/ClaimId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimDocumentSubmission"
      responses:
        "201":
          description: The attachment as the systems recorded it.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimDocument"
        "404":
          $ref: "#/components/responses/ClaimNotFound"
        "409":
          $ref: "#/components/responses/IdempotencyKeyReuse"
        "422":
          $ref: "#/components/responses/InvalidPayload"
  /claims/{claim_id}/parties:
    get:
      operationId: read_claim_parties
      summary: Read the parties a claim has so far
      description: |
        The order is the order the parties were recorded in, oldest first.
      parameters:
        - $ref: "#/components/parameters/ClaimId"
      responses:
        "200":
          description: Every party recorded against the claim.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClaimParty"
        "404":
          $ref: "#/components/responses/ClaimNotFound"
        "422":
          $ref: "#/components/responses/InvalidPayload"
    post:
      operationId: record_claim_party
      summary: Record a party the pipeline identified
      description: |
        Party history is a history because something appends to it: the pipeline records each party as
        extraction identifies it (v0.2), and this is where it does so. Without the append, the read
        above could only ever answer with an empty list.
      parameters:
        - $ref: "#/components/parameters/ClaimId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimPartySubmission"
      responses:
        "201":
          description: The party as the systems recorded it.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimParty"
        "404":
          $ref: "#/components/responses/ClaimNotFound"
        "409":
          $ref: "#/components/responses/IdempotencyKeyReuse"
        "422":
          $ref: "#/components/responses/InvalidPayload"
components:
  parameters:
    ClaimId:
      name: claim_id
      in: path
      required: true
      description: The claim the surrounding systems hold.
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        The key this write is deduplicated by: a retry carrying the key and the same request answers
        with the response the first call recorded and changes nothing. Reusing the key for a
        different request is refused as `idempotency_key_reuse`.
      schema:
        type: string
        minLength: 1
        maxLength: 128
  responses:
    ClaimNotFound:
      description: The systems hold no such claim.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    PolicyNotFound:
      description: The systems hold no such policy.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    IdempotencyKeyReuse:
      description: The key was already used for a different request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InvalidPayload:
      description: The request is outside the contract, or the required header is missing.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServiceUnavailable:
      description: The systems are up but the claims they hold are not reachable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorCode:
      type: string
      description: What the systems refused a call for.
      # OpenAPI has no per-member description; `x-enum-descriptions` is the convention this document
      # and the generator share, and it is what the generated client's typed errors document.
      x-enum-descriptions:
        invalid_payload: The request is outside the contract, or a required header is missing.
        claim_not_found: The systems hold no such claim.
        policy_not_found: The systems hold no such policy.
        idempotency_key_reuse: The key was already used for a different request.
        service_unavailable: The systems are up, but the claims they hold are not reachable.
      enum:
        - invalid_payload
        - claim_not_found
        - policy_not_found
        - idempotency_key_reuse
        - service_unavailable
    ErrorResponse:
      type: object
      additionalProperties: false
      description: The shape of every failure the systems answer with.
      required: [code, detail]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        detail:
          type: string
          description: What exactly was wrong, naming the field or the key that caused it.
    Readiness:
      type: object
      additionalProperties: false
      description: What a ready system of record answers with.
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]
    Policy:
      type: object
      additionalProperties: false
      description: A policy as the system of record holds it.
      required: [policy_number, product_family, valid_from]
      properties:
        policy_number:
          type: string
          description: The number the claim is made against.
        product_family:
          type: string
          description: The product line the policy belongs to, e.g. `Auto & pohoda`.
        valid_from:
          type: string
          format: date
          description: The day cover started.
    ClaimStatus:
      type: string
      description: |
        Where the claim stands as the surrounding systems see it. v0.1 has two positions: `received`
        for a claim the systems have taken in, and `triaged` for one the pipeline has finished with.
        The routing queues and human-review states of increment v0.3 extend this.
      enum:
        - received
        - triaged
    ClaimStatusTransition:
      type: object
      additionalProperties: false
      description: One status the claim has held, and when it started holding it.
      required: [status, recorded_at]
      properties:
        status:
          $ref: "#/components/schemas/ClaimStatus"
        recorded_at:
          type: string
          format: date-time
    ClaimSubmission:
      type: object
      additionalProperties: false
      description: A claim as the surrounding systems accept it.
      required: [policy_number, incident_date, claim_amount_eur]
      properties:
        policy_number:
          type: string
          minLength: 1
          maxLength: 64
        incident_date:
          type: string
          format: date
          description: The day the loss happened, which decides the edition a clause is read from.
        claim_amount_eur:
          $ref: "#/components/schemas/EurAmount"
    ClaimStatusUpdate:
      type: object
      additionalProperties: false
      description: The status the pipeline reports back to the surrounding systems.
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/ClaimStatus"
    Claim:
      type: object
      additionalProperties: false
      description: A stored claim, the record the systems return for every read.
      required: [claim_id, policy_number, incident_date, claim_amount_eur, status, status_history]
      properties:
        claim_id:
          type: string
          format: uuid
        policy_number:
          type: string
        incident_date:
          type: string
          format: date
        claim_amount_eur:
          $ref: "#/components/schemas/EurAmount"
        status:
          $ref: "#/components/schemas/ClaimStatus"
        status_history:
          type: array
          minItems: 1
          description: Every status this claim has held, oldest first; `status` is the last of them.
          items:
            $ref: "#/components/schemas/ClaimStatusTransition"
    EurAmount:
      type: string
      format: decimal
      exclusiveMinimum: 0
      x-max-digits: 12
      x-decimal-places: 2
      description: An amount in euro, at most two decimal places, as a decimal string.
    ClaimDocumentSubmission:
      type: object
      additionalProperties: false
      description: A document the pipeline read out of a claim, as it attaches it.
      required: [document_type, filename]
      properties:
        document_type:
          type: string
          minLength: 1
          maxLength: 64
          description: |
            What the pipeline classified the document as. v0.2's schema registry owns this
            vocabulary; until then it is free text.
        filename:
          type: string
          minLength: 1
          maxLength: 255
    ClaimDocument:
      type: object
      additionalProperties: false
      description: A document the surrounding systems hold against a claim.
      required: [document_id, claim_id, document_type, filename, attached_at]
      properties:
        document_id:
          type: string
          format: uuid
        claim_id:
          type: string
          format: uuid
        document_type:
          type: string
        filename:
          type: string
        attached_at:
          type: string
          format: date-time
    PartyRole:
      type: string
      description: What a party is to the claim.
      enum:
        - policyholder
        - claimant
        - driver
        - witness
        - other
    ClaimPartySubmission:
      type: object
      additionalProperties: false
      description: A party the pipeline identified, as it records it.
      required: [role]
      properties:
        role:
          $ref: "#/components/schemas/PartyRole"
    ClaimParty:
      type: object
      additionalProperties: false
      description: A party the surrounding systems hold against a claim.
      required: [party_id, claim_id, role, recorded_at]
      properties:
        party_id:
          type: string
          format: uuid
          description: The systems' own identifier; the pipeline's identification of the party stays with the pipeline.
        claim_id:
          type: string
          format: uuid
        role:
          $ref: "#/components/schemas/PartyRole"
        recorded_at:
          type: string
          format: date-time